feat/redux-store_active-thread: Extract props drawerState, setDrawerState and toggleDrawer from HeadNav to Main
wip
